backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.24.0
pydantic>=2.7.0
pydantic-settings>=2.3.1
python-dotenv>=1.0.1
openai>=1.33.0,<2
chromadb>=0.5.4,<0.6
numpy>=1.26.0
